refactor(tasks): move archivable status logic to utils and clean up tests
Removed the isTaskArchivableStatus function and related constants from the task helper, centralizing the logic in the utils package. Updated tests to reflect these changes and ensure proper imports. This refactor aims to streamline task status management across the application.
