+++ apps/core/src/routes/v1/tasks/[id]/delete.ts
@@ -1,14 +1,15 @@
 import { createRoute, z } from "@hono/zod-openapi";
 
+import {
+  getTaskCannotArchiveMessage,
+  isTaskArchivableStatus,
+} from "@sokosumi/utils";
+
 import { requireTaskOwnership } from "@/helpers/access-control";
 import { unprocessableEntity } from "@/helpers/error";
 import { jsonErrorResponse, jsonSuccessResponse } from "@/helpers/openapi";
 import { ok } from "@/helpers/response";
-import {
-  getTaskCannotArchiveMessage,
-  isTaskArchivableStatus,
-  mapTask,
-} from "@/helpers/task";
+import { mapTask } from "@/helpers/task";
 import prisma from "@/lib/db/prisma";
 import type { OpenAPIHonoWithAuth } from "@/lib/hono";
 import { requireUserContext } from "@/middleware/auth";
